+++ cluster_tools/Changelog.md
@@ -14,9 +14,10 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Added
 
 ### Changed
+- Stop 
 
 ### Fixed
-
+- Fixed `forward_log` raising an unhandled `TailError` when a cluster job completes without a log file being visible from the head node (e.g. due to slow NFS). [#1480](https://github.com/scalableminds/webknossos-libs/pull/1480)
 
 ## [3.5.2](https://github.com/scalableminds/webknossos-libs/releases/tag/v3.5.2) - 2026-06-18
 [Commits](https://github.com/scalableminds/webknossos-libs/compare/v3.5.1...v3.5.2)